- Added to feature set - Disabled auto opening - Added toggle in main list to switch patch state - Moved IL to seperate class - All windows are now draggable and resizable - Proper scaling - Fixed Typos & Logging - Fix collection modify error
